pyproject.toml
[project]
name = "sigk"
version = "0.1.0"
description = "sigk_g10"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "lpips",
    "scikit-image>=0.21.0",
    "requests>=2.28.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
